monitor/requirements.txt
stomp.py==8.1.0
python-dotenv==1.0.1
